refactor: extract map runner static deadlock helpers
